+++ math_expr/README.md
@@ -95,3 +95,42 @@ We evaluate a program on a test graph _t_ as follows:
    nothing or painting everything black.
 
 
+## Hypothesis Formation and Evaluation
+
+The task of a _program synthesiser_ is to receive as input the bitmap
+of a graph and output a program in this programming language. For the
+purposes of this exercise the synthesier may assume that all inputs
+are 640x480 bitmaps, centered on (0,0), without axes or any other
+decorations, and that the background is white and the graph is
+non-white.
+
+The hypothesis generated by the program synthesier is evaluated by
+first using the code in `executor.py` to produce a bitmap and the
+using the code in `evaluator.py` to evaluate the similarity between
+the input bitmap and the bitmap generated by executing the hypothesis.
+In other words, there is no "golden truth" program, but rather a
+target bitmap and any program that matches it is evaluated positively.
+
+It should be noted that although some effort has been made to reduce
+the different syntact variations for expressing identical semantics,
+this is not completely possible. At the very least, any poly can be
+broken up into multiple lines. Furthermore, the graphs are evaluated
+on a specific domain, thre will be semantically different functions
+that are indistiguishable within this domain.
+
+
+## Hypothesis Formation and Evaluation
+
+The task of the exercise is to create a program synthesiser that is
+trained on examples of programs and target bitmaps, _without being
+given access to the basic mathematical functions_. The program
+synthesiser will be made aware of the operators of the language and
+how they are used to combine the outputs of the different functions,
+but not of the functions themselves.
+
+Think of this as a computer vision exercise, where the machine
+learning will learn how to recognize the elemntary functions
+(sine, sinc, exponential, logarithmic) although it is only given
+as supervision descriptions of complex bitmaps (e.g., this is a
+composition of exp and sine) without ever observing the elementary
+functions in isolation.
