+++ math_expr/graphs/exp.graph
@@ -1,0 +1,1 @@
+exp( )
