+++ math_expr/graphs/exp_in_sine.graph
@@ -1,0 +1,1 @@
+sin(), poly([0,20]) , exp()
