+++ math_expr/graphs/exp_in_sine_in_poly.graph
@@ -1,0 +1,1 @@
+poly([0,0.1,0.3,0.5]),sin(),poly([0,20]),exp ()
