+++ math_expr/graphs/exp_in_sine_on_poly.graph
@@ -1,0 +1,2 @@
+sin(),poly([0,20]),exp()
+poly([0,0.1,0.3,0.5])
