+++ math_expr/graphs/natlog.graph
@@ -1,0 +1,1 @@
+ln ()
