+++ math_expr/graphs/poly.graph
@@ -1,0 +1,1 @@
+poly([0,0.001,0.003,0.002])
