+++ math_expr/graphs/sin_in_ln.graph
@@ -1,0 +1,1 @@
+ln(),sin()
