+++ math_expr/graphs/sin_poly.graph
@@ -1,0 +1,2 @@
+poly([0,0.1,0.3,0.2]) 
+poly([0,5]),sin(),poly([0,10])
