+++ math_expr/graphs/sinc.graph
@@ -1,0 +1,1 @@
+sinc ()
