+++ math_expr/graphs/sinc_linear.graph
@@ -1,0 +1,2 @@
+sinc() , poly( [2,   1]    )
+         poly( [0, 0.2] )
