+++ math_expr/graphs/sine.graph
@@ -1,0 +1,1 @@
+poly( [0,2]) , sin() ,poly([0, 3])
