+++ math_expr/graphs/sine_in_exp.graph
@@ -1,0 +1,1 @@
+exp(),sin()
